+++ src/main/java/duke/Storage.java
@@ -1,0 +1,4 @@
+package duke;// deals with loading tasks from the file and saving tasks in the file
+
+public class Storage {
+}
