+++ src/main/java/duke/tasks/Task.java
@@ -1,0 +1,33 @@
+package duke.tasks;
+
+// This class represents the tasks added by the user
+public class Task {
+    protected final String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public Task(String description, boolean isDone) {
+        this.description = description;
+        this.isDone = isDone;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done task with X
+    }
+
+    public void markAsDone() {
+        this.isDone = true;
+    }
+
+    public String toString() {
+        return "[" + this.getStatusIcon() + "] ";
+    }
+
+    public String getStatusString() { return ""; }
+
+}
+
